+++ callbacks.py
@@ -1,0 +1,15 @@
+from dash.dependencies import Input, Output
+
+from app import app
+
+@app.callback(
+    Output('app-1-display-value', 'children'),
+    [Input('app-1-dropdown', 'value')])
+def display_value(value):
+    return 'You have selected "{}"'.format(value)
+
+@app.callback(
+    Output('app-2-display-value', 'children'),
+    [Input('app-2-dropdown', 'value')])
+def display_value(value):
+    return 'You have selected "{}"'.format(value)
