+++ ledger_explorer.py
@@ -57,103 +57,6 @@ def get_descendents(account_id, account_tree):
     return [x.tag for x in descendent_nodes]
 
 
-def load_eras(source, earliest_date, latest_date):
-    """
-    If era data file is available, use it to construct
-    arbitrary bins
-    """
-
-    try:
-        data = pd.read_csv(source)
-    except urllib.error.HTTPError:
-        return None
-
-    data = data.astype({'start_date': 'datetime64'})
-    data = data.astype({'end_date': 'datetime64'})
-
-    data = data.sort_values(by=['start_date'], ascending=False)
-    data = data.reset_index(drop=True).set_index('name')
-
-    if pd.isnull(data.iloc[0].end_date):
-        data.iloc[0].end_date = latest_date
-    if pd.isnull(data.iloc[-1].start_date):
-        data.iloc[-1].start_date = earliest_date
-
-    return data
-
-
-def load_transactions(source):
-    """
-    Load a csv matching the transaction export format from Gnucash.
-    Uses columns 'Account Name', 'Description', 'Memo', Notes', 'Full Account Name', 'Date', 'Amount Num.'
-    """
-
-    def convert(s):  # not fast
-        dates = {date: pd.to_datetime(date) for date in s.unique()}
-        return s.map(dates)
-    data = pd.read_csv(source, thousands=',')
-    data.columns = [x.lower() for x in data.columns]
-    data['date'] = data['date'].astype({'date': 'datetime64'})
-
-    # Gnucash doesn't include the date, description, or notes for transaction splits.  Fill them in.
-    data['date'] = data['date'].fillna(method='ffill')
-    data['description'] = data['description'].fillna(method='ffill')
-    data['notes'] = data['notes'].fillna(method='ffill')
-
-    # data['date'] = convert(data['date']) # supposedly faster, but not actually much faster, and broken
-    data = data.rename(columns={'amount num.': 'amount', 'account name': 'account'})
-
-    data['amount'] = data['amount'].replace(to_replace=',', value='')
-    data['amount'] = data['amount'].astype(float).round(decimals=0).astype(int)
-
-    data.fillna('', inplace=True)  # Any remaining fields with invalid numerical data should be text fields
-    data.where(data.notnull(), None)
-
-    data['memo'] = data['memo'].astype(str)
-    data['description'] = data['description'].astype(str)
-    data['notes'] = data['notes'].astype(str)
-
-    data['description'] = (data['description'] + ' ' + data['memo'] + ' ' + data['notes']).str.strip()
-    trans = data[['date', 'description', 'amount', 'account', 'full account name']]
-    account_tree = make_account_tree_from_trans(trans)
-    return trans, account_tree
-
-
-def make_account_tree_from_trans(trans):
-    """ extract all accounts from a list of Gnucash account paths
-
-    Each account name is a full path.  Parent accounts with no
-    transactions will be missing from the data, so reconstruct the
-    complete tree implied by the transaction data.
-
-    As long as the accounts are sorted hierarchically, the algorithm
-    should never encounter a missing parent except the first node.
-
-    If there are multiple heads in the data, they will all belong to
-    root, so the tree will still be a DAG
-    """
-
-    tree = treelib.Tree()
-    tree.create_node(tag='All', identifier='root')
-    accounts = trans['full account name'].unique()
-
-    for account in accounts:
-        branches = account.split(':')  # example: Foo:Bar:Baz
-        for i, branch in enumerate(branches):
-            name = branch
-            if i == 0:
-                parent = 'root'
-            else:
-                parent = branches[i-1]
-            if not tree.get_node(name):
-                tree.create_node(tag=name,
-                                 identifier=name,
-                                 parent=parent)
-
-    tree = trim_excess_root(tree)
-    return tree
-
-
 def make_bar(account, color_num=0, time_resolution=0, time_span=1, deep=False):
     """ returns a go.Bar object with total by time_resolution period for
     the selected account.  If deep, include total for all descendent accounts. """
@@ -490,18 +393,6 @@ def positize(trans):
     return trans
 
 
-def trim_excess_root(tree):
-    # If the input tree's root has no branches, trim the superfluous node and return a shorter tree
-    root_id = tree.root
-    root_kids = tree.children(root_id)
-    if len(root_kids) == 1:
-        tree.update_node(root_kids[0].identifier, parent=None, bpointer=None)
-        new_tree = tree.subtree(root_kids[0].identifier)
-        return new_tree
-    else:
-        return tree
-
-
 #######################################################################
 # Initialize and set up formatting
 #######################################################################
@@ -519,394 +410,16 @@ app.css.append_css(dict(external_url='http://localhost/dash_layout.css'))
 
 pd.set_option('display.max_rows', None)  # useful for DEBUGging, put back to 10?
 
-disc_colors = px.colors.qualitative.D3
-
-big_font = dict(
-    family='IBM Plex Sans Medium',
-    size=24)
-
-medium_font = dict(
-    family='IBM Plex Sans Light',
-    size=20)
-
-small_font = dict(
-    family='IBM Plex Light',
-    size=12)
-
-time_series_layout = dict(
-    legend={'x': 0, 'y': 1},
-    font=small_font,
-    paper_bgcolor='var(--bg)',
-    titlefont=medium_font)
-
-
-chart_fig_layout = dict(
-    clickmode='event+select',
-    dragmode='select',
-    margin=dict(
-        l=10,  # NOQA
-        r=10,
-        t=10,
-        b=10),
-    height=350,
-    paper_bgcolor='var(--bg)',
-    showlegend=False,
-    title=dict(
-            font=big_font,
-            x=0.1,
-            y=0.9),
-    hoverlabel=dict(
-        bgcolor='var(--bg)',
-        font_color='var(--fg)',
-        font=medium_font))
-
-TIME_RES_LOOKUP = {
-    0: {'label': 'All', 'abbrev': 'all'},
-    1: {'label': 'Era', 'abbrev': 'era'},
-    2: {'label': 'Year', 'abbrev': 'Y', 'resample_keyword': 'A', 'months': 12},
-    3: {'label': 'Quarter', 'abbrev': 'Q', 'resample_keyword': 'Q', 'months': 3},
-    4: {'label': 'Month', 'abbrev': 'Mo', 'resample_keyword': 'M', 'months': 1}}
-
-TIME_RES_OPTIONS = {key: value['label'] for key, value in TIME_RES_LOOKUP.items()}
-
-TIME_SPAN_LOOKUP = {
-    0: {'label': 'Annual', 'abbrev': ' ⁄y', 'months': 12},
-    1: {'label': 'Monthly', 'abbrev': ' ⁄mo', 'months': 1}}
-
-TIME_SPAN_OPTIONS = {key: value['label'] for key, value in TIME_SPAN_LOOKUP.items()}
-
-#######################################################################
-# Load Data
-#######################################################################
-
-# this could come from a URL; simpler now to get from a local file
-# crash if the load fails, as nothing is going to work
-
-trans, account_tree = load_transactions('http://localhost/transactions.csv')
-trans = trans.sort_values(['date', 'account'])
-earliest_trans = trans['date'].min()
-latest_trans = trans['date'].max()
-
-# Load a custom eras file if present.
-#
-eras = load_eras('http://localhost/eras.csv', earliest_trans, latest_trans)
-SUBTOTAL_SUFFIX = ' Subtotal'
-LEAF_SUFFIX = ' Leaf'
-OTHER_PREFIX = 'Other '
-MAX_SLICES = 7  # TODO: expose this in a control
 
 
 #######################################################################
 # Declare the content of the page
 #######################################################################
-trans_table = dash_table.DataTable(
-    id='trans_table',
-    columns=[dict(id='date', name='Date'),
-             dict(id='account', name='Account'),
-             dict(id='description', name='Description'),
-             dict(id='amount', name='Amount')],
-    style_header={'font-family': 'IBM Plex Sans, Verdana, sans',
-                  'font-size=': '1.1rem',
-                  'text-align': 'center'},
-    style_cell={'overflow': 'hidden',
-                'textOverflow': 'ellipsis',
-                'maxWidth': 0,
-                'backgroundColor': 'var(--bg-more)'},
-    style_cell_conditional=[
-        {'if': {'column_id': 'date'},
-         'textAlign': 'left',
-         'padding': '0px 10px',
-         'width': '18%'},
-        {'if': {'column_id': 'account'},
-         'textAlign': 'left',
-         'padding': '0px px',
-         'width': '20%'},
-        {'if': {'column_id': 'description'},
-         'textAlign': 'left',
-         'padding': 'px 2px 0px 3px'},
-        {'if': {'column_id': 'amount'},
-         'padding': '0px 12px 0px 0px',
-         'width': '13%'}],
-    data=[],
-    sort_action='native',
-    page_action='native',
-    filter_action='native',
-    style_as_list_view=True,
-    page_size=20)
-app.layout = html.Div(
-    className="layout_box",
-    children=[
-        html.Div(
-            id='time_series_control_bar',
-            className="control_bar dashbox",
-            children=[
-                dcc.Slider(
-                    className='resolution-slider',
-                    id='time_series_resolution',
-                    min=0,
-                    max=4,
-                    step=1,
-                    marks=TIME_RES_OPTIONS,
-                    value=1
-                ),
-                dcc.Slider(
-                    className='span-slider',
-                    id='time_series_span',
-                    min=0,
-                    max=1,
-                    step=1,
-                    marks=TIME_SPAN_OPTIONS,
-                    value=1
-                )
-            ]),
-        html.Div(
-            id='detail_control_bar',
-            className="control_bar dashbox",
-            children=[
-                html.H2(
-                    id='selected_account_display',
-                    children=['Account']),
-                html.H2(
-                    id='burst_selected_account_display',
-                    children=[]),
-                html.H2(
-                    id='selected_date_range_display',
-                    children=['All Dates']),
-                dcc.Store(id='detail_store',
-                          storage_type='memory')
-            ]),
-        html.Div(
-            className='account_burst dashbox',
-            children=[
-                dcc.Graph(
-                    id='account_burst',
-                    figure=make_sunburst(trans, SUBTOTAL_SUFFIX=SUBTOTAL_SUFFIX))
-            ]),
-        html.Div(
-            className='master_time_series dashbox',
-            children=[
-                dcc.Graph(
-                    id='master_time_series')
-            ]),
-        html.Div(
-            className="trans_table dashbox",
-            children=[
-                trans_table
-            ]),
-        html.Div(
-            className='transaction_time_series dashbox',
-            children=[
-                dcc.Graph(
-                    id='transaction_time_series')
-            ]),
-    ])
 #######################################################################
 # Callback functions
 #######################################################################
 
 
-@app.callback(
-    [Output('master_time_series', 'figure')],
-    [Input('time_series_resolution', 'value'),
-     Input('time_series_span', 'value')])
-def apply_time_series_resolution(time_resolution, time_span):
-    try:
-        tr = TIME_RES_LOOKUP[time_resolution]
-        ts = TIME_SPAN_LOOKUP[time_span]
-        ts_label = ts.get('label')      # e.g., 'Annual' or 'Monthly'
-        tr_label = tr.get('label')          # e.g., 'by Era'
-    except IndexError:
-        logging.critical(f'Bad data from period selectors: time_resolution {time_resolution}, time_span {time_span}')
-        return
-
-    chart_fig = go.Figure(layout=chart_fig_layout)
-    root_account_id = account_tree.root  # TODO: Stub for controllable design
-    selected_accounts = get_children(root_account_id, account_tree)
-
-    for i, account in enumerate(selected_accounts):
-        chart_fig.add_trace(make_bar(account, i, time_resolution, time_span, deep=True))
-
-    chart_fig.update_layout(
-        title={'text': f'Average {ts_label} $, by {tr_label} '},
-        xaxis={'showgrid': True, 'dtick': 'M3'},
-        barmode='relative')
-
-    return [chart_fig]
-
-
-@app.callback(
-    [Output('selected_account_display', 'children'),
-     Output('selected_date_range_display', 'children'),
-     Output('detail_store', 'data'),
-     Output('account_burst', 'figure')],
-    [Input('master_time_series', 'figure'),
-     Input('master_time_series', 'selectedData')])
-def apply_selection_from_time_series(figure, selectedData):
-    """
-    Selecting specific points from the time series chart updates the
-    account burst and the detail labels.
-
-    Reminder to self: When you think selectedData input is broken, remember
-    that unaltered default action in the graph is to zoom, not to select.
-
-    Note: all of the necessary information is in figure but that doesn't seem
-    to trigger reliably.  Adding selectedData as a second Input causes reliable
-    triggering.
-
-    TODO: maybe check for input safety?
-
-    """
-    selection_start_date = None
-    selection_end_date = None
-    date_range_content = None
-    filtered_trans = None
-    selected_accounts = []
-    detail_store = None
-
-    for trace in figure.get('data'):
-        account = trace.get('name')
-        points = trace.get('selectedpoints')
-        if not points:
-            continue
-        selected_accounts.append(account)
-        for point in points:
-            # back out the selection parameters (account and start/end dates)
-            # from the trace
-            # TODO: for All, x is end date
-            #       for By Era, x is start date
-            #       for A/Q/M, x is end date
-            # so fix that.
-            selection_start_date = pd.to_datetime(trace['x'][point])
-            # the last point in the time-series won't have a following point
-            try:
-                selection_end_date = pd.to_datetime(trace['x'][point + 1])
-            except IndexError:
-                selection_end_date = latest_trans
-
-            logging.debug(f'point: start {selection_start_date}, end {selection_end_date}')
-            point_accounts = get_descendents(account, account_tree)
-
-            new_trans = trans.loc[trans['account'].isin(point_accounts)].\
-                loc[trans['date'] >= selection_start_date].\
-                loc[trans['date'] <= selection_end_date]
-
-            try:
-                filtered_trans.append(new_trans)
-            except AttributeError:
-                filtered_trans = new_trans
-
-    # If no transactions are ultimately selected, show all transactions
-    try:
-        data_count = len(filtered_trans)
-    except TypeError:
-        data_count = 0
-    if data_count == 0:
-        filtered_trans = trans
-        selected_accounts = ['All']
-
-    pos_trans = positize(filtered_trans)
-    sun_fig = make_sunburst(pos_trans, selection_start_date, selection_end_date, SUBTOTAL_SUFFIX=SUBTOTAL_SUFFIX)
-    account_children = ', '.join(selected_accounts)
-    if selection_start_date and selection_end_date:
-        date_range_content = ['Between ',
-                              selection_start_date.strftime("%Y-%m-%d"),
-                              ' and ',
-                              selection_end_date.strftime("%Y-%m-%d")]
-        detail_store = {'start': selection_start_date, 'end': selection_end_date}
-    return [account_children, date_range_content, detail_store, sun_fig]
-
-
-@app.callback(
-    [Output('trans_table', 'data'),
-     Output('transaction_time_series', 'figure'),
-     Output('burst_selected_account_display', 'children')],
-    [Input('account_burst', 'clickData'),
-     Input('detail_store', 'data')])
-def apply_burst_click(burst_clickData, detail_data):
-    """
-    Clicking on a slice in the Sunburst updates the transaction list with matching transactions
-
-    TODO: maybe check for input safety?
-    """
-    selected_accounts = []
-    tts_fig = go.Figure(layout=chart_fig_layout)
-
-    if burst_clickData:
-        click_account = burst_clickData['points'][0]['id']
-    else:
-        click_account = []
-
-    if click_account:
-        try:
-            selected_accounts = [click_account] + get_children(click_account, account_tree)
-        except treelib.exceptions.NodeIDAbsentError:
-            # This is a hack.  If the account isn't there, assume that the reason
-            # is that it was reidentified to 'X Leaf', and back that out.
-            try:
-                if LEAF_SUFFIX in click_account:
-                    revised_id = click_account.replace(LEAF_SUFFIX, '')
-                    selected_accounts = [revised_id]
-            except treelib.exceptions.NodeIDAbsentError:
-                pass
-    else:
-        title = 'All'
-        pass
-
-    if selected_accounts:
-        title = click_account
-        sel_trans = trans[trans['account'].isin(selected_accounts)]
-    else:
-        sel_trans = trans
-
-    try:
-        start_date = detail_data['start']
-        end_date = detail_data['end']
-        sel_trans = sel_trans[(sel_trans['date'] >= start_date) & (sel_trans['date'] <= end_date)]
-        tts_fig.add_shape(
-            type="rect",
-            x0=earliest_trans,
-            x1=start_date,
-            yref='paper',
-            ysizemode='scaled',
-            y0=0,
-            y1=1,
-            line=dict(width=0),
-            fillcolor='#e1e1e1',
-            opacity=0.6
-        )
-        tts_fig.add_shape(
-            type="rect",
-            x0=end_date,
-            x1=latest_trans,
-            yref='paper',
-            ysizemode='scaled',
-            y0=0,
-            y1=1,
-            line=dict(width=0),
-            fillcolor='#e1e1e1',
-            opacity=0.6
-        )
-
-    except (KeyError, TypeError):
-        pass
-
-    if len(selected_accounts) == 1:
-        try:
-            account = selected_accounts[0]
-            tts_fig.add_trace(make_scatter(account, sel_trans))
-        except TypeError:
-            pass
-    elif len(selected_accounts) > 1:
-        # TODO: design bug: certain leaves (maybe with ¿hidden children?)
-        # appear as leaves in the sunburst and show only one
-        # account in the trans table, but have multiple selected_accounts
-        # and so render as bars when they should be scatter
-        for i, account in enumerate(selected_accounts):
-            tts_fig.add_trace(make_bar(account, i, 4, 1, deep=True))
-            tts_fig.update_layout(
-                barmode='stack',
-                showlegend=True)
-    return [sel_trans.to_dict('records'), tts_fig, title]
 
 
 if __name__ == '__main__':
